+++ src/frequenz/sdk/timeseries/battery_pool/__init__.py
@@ -3,11 +3,11 @@
 
 """Manage a pool of batteries."""
 
-from ._result_types import Bound, PowerMetrics
+from ._result_types import Bounds, PowerMetrics
 from .battery_pool import BatteryPool
 
 __all__ = [
     "BatteryPool",
     "PowerMetrics",
-    "Bound",
+    "Bounds",
 ]